Move testcase into LogicalPlanToSQLSuite.
