--- sql/hive/src/test/scala/org/apache/spark/sql/hive/SQLBuilderSuite.scala
@@ -1,33 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.hive
-
-import org.apache.spark.sql.QueryTest
-import org.apache.spark.sql.catalyst.SQLBuilder
-import org.apache.spark.sql.hive.test.TestHiveSingleton
-import org.apache.spark.sql.test.SQLTestUtils
-
-class SQLBuilderSuite extends QueryTest with SQLTestUtils with TestHiveSingleton {
-  test("predicate subquery") {
-    withTable("t1") {
-      sql("CREATE TABLE t1(a int)")
-      val df = sql("select * from t1 b where exists (select * from t1 a)")
-      new SQLBuilder(df).toSQL
-    }
-  }
-}
